backend/requirements.txt
django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
boto3==1.34.84
django-storages==1.14.2
python-magic==0.4.27